test(fn_test): divide in create/invoke/delete test files
